+++ src/proof_verifier.cpp
@@ -1,0 +1,13 @@
+// Copyright (c) 2016-2020 The Zcash developers
+// Distributed under the MIT software license, see the accompanying
+// file COPYING or https://www.opensource.org/licenses/mit-license.php .
+
+#include <proof_verifier.h>
+
+ProofVerifier ProofVerifier::Strict() {
+    return ProofVerifier(true);
+}
+
+ProofVerifier ProofVerifier::Disabled() {
+    return ProofVerifier(false);
+}
