+++ src/test/checkblock_tests.cpp
@@ -5,6 +5,7 @@
 #include "clientversion.h"
 #include "consensus/validation.h"
 #include "main.h"
+#include "proof_verifier.h"
 #include "test/test_bitcoin.h"
 #include "utiltime.h"
 #include "zcash/Proof.hpp"
@@ -57,7 +58,7 @@ BOOST_AUTO_TEST_CASE(May15)
 
         // After May 15'th, big blocks are OK:
         forkingBlock.nTime = tMay15; // Invalidates PoW
-        auto verifier = libzcash::ProofVerifier::Strict();
+        auto verifier = ProofVerifier::Strict();
         BOOST_CHECK(CheckBlock(forkingBlock, state, Params(), verifier, false, false));
     }
