+++ src/zcash/Proof.hpp
@@ -211,29 +211,6 @@ typedef std::array<unsigned char, GROTH_PROOF_SIZE> GrothProof;
 // with a PHGRProof. The default however should be GrothProof.
 typedef boost::variant<PHGRProof, GrothProof> SproutProof;
 
-class ProofVerifier {
-private:
-    bool perform_verification;
-
-    ProofVerifier(bool perform_verification) : perform_verification(perform_verification) { }
-
-public:
-    // ProofVerifier should never be copied
-    ProofVerifier(const ProofVerifier&) = delete;
-    ProofVerifier& operator=(const ProofVerifier&) = delete;
-    ProofVerifier(ProofVerifier&&);
-    ProofVerifier& operator=(ProofVerifier&&);
-
-    // Creates a verification context that strictly verifies
-    // all proofs.
-    static ProofVerifier Strict();
-
-    // Creates a verification context that performs no
-    // verification, used when avoiding duplicate effort
-    // such as during reindexing.
-    static ProofVerifier Disabled();
-};
-
 }
 
 #endif // ZC_PROOF_H_
